backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
paramiko==3.5.1
cryptography==44.0.0
jinja2==3.1.5
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
aiofiles==24.1.0
ansible-core==2.16.3
httpx==0.27.0
bcrypt==5.0.0
PyJWT==2.11.0
